Fixed listeners
This was mainly a fix for the listeners. The previous commit didn't allow the user to select and recalculate the correct settings. This now let's the user click-and-drag a box around the region they want to look closer at.

Additionally, it now gets rid of all the previously known points to help speed up the graph, which slows down immensely when there's a large number of points.