Refactoring for bordering point calculations.
This was overall a major simplification to the original way of finding bordering points and then finding points past that.

There's still some more work to do with this, but this does add resolution and help show the finer points of the fractal.

The work needed to still do with this is that the program sometimes doesn't realize when it is working with a border point, which is evidenced by the gaps around the edges of the graph currently.