Refactored point calculator
The previous way of finding points would be to look around for patterns in the points. If there was a pattern, that'd mean the point was in. This was slow and required the program to constantly be looking back through arrays.

The new method only looks at the magnitude of the complex numbers being used. This doesn't require we constantly through all the previous data associated with a point, which means the program runs faster.

In addition, this greatly simplified the code so the math isn't terribly convoluted.