Optimization
Points that come successive to points already found to be in the set now have a lower barrier of entry to being in the set.

This is because points that are in the set are usually near to points also in the set, so if several points in a row are all in the set, then the point after those is probably in there too, so we shouldn't look so closely at it.

The tweak essentially is corner-cutting, but it doesn't cause any significant errors while speeding along the graphing process a fair bit.