neighboring point functionality for border points
The neighboring function is being used again and was refactored. The program no longer needs to search through the entire array to find where border points lie. It finds those as the array is calculated and only finds neighbors of border points.

A border point is any point at the start of a vertical line of points that are in the set, the prior in-set point of any out-of-set points, and the first in-set after any number of out-of-set points.

Additionally, a few more variables have been added that are console.logged after the graph's completion. These are to keep track of the program and give some data on things like the number of points that are in the set and how many neighboring points have been found.